+++ fist-kit-infra/fist-trace-context-spring-boot-starter/src/main/java/com/power4j/fist/trace/boot/autoconfigure/TraceContextServletFilter.java
@@ -6,6 +6,8 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
@@ -16,6 +18,7 @@ import java.io.IOException;
  * @author CJ (power4j@outlook.com)
  * @since 3.15
  */
+@Order(Ordered.HIGHEST_PRECEDENCE)
 public class TraceContextServletFilter extends OncePerRequestFilter {
 
 	private final TraceContextRuntime runtime;